Bug #0000 [FE] Intro coach screen is missing for the user to navigate between videos, Bug #0000 [FE] Questions button has a drop shadow that is missing in the developed version (#139)
* [FE] Intro coach screen is missing for the user to navigate between videos

* fix: shadow to video

---------
