+++ src/components/common/icons/Icon.tsx
@@ -288,6 +288,52 @@ const iconsMap: Record<string, React.FC<IconProps>> = {
       </Icon>
     );
   },
+  ScrollHandIcon: ({ color, active, ...props }) => {
+    return (
+      <Icon viewBox="0 0 42 59" color={color} {...props}>
+        <mask
+          id="mask0_1643_6120"
+          maskUnits="userSpaceOnUse"
+          x="0"
+          y="0"
+          width="48"
+          height="48"
+        >
+          <rect
+            y="21.8547"
+            width="34"
+            height="34"
+            transform="rotate(-40 0 21.8547)"
+            fill="#D9D9D9"
+          />
+        </mask>
+        <g mask="url(#mask0_1643_6120)">
+          <path
+            d="M29.921 35.5837C29.4326 35.9935 28.8796 36.2803 28.262 36.4441C27.6443 36.608 27.0224 36.629 26.3962 36.5072L12.1071 33.7487L12.2434 32.4785C12.2987 31.8465 12.5387 31.3061 12.9634 30.8572C13.3882 30.4083 13.9043 30.1602 14.5118 30.1128L19.5029 29.6696L10.3285 18.7359C10.0705 18.4284 9.95821 18.0834 9.99167 17.7009C10.0251 17.3183 10.1956 16.9981 10.5031 16.7401C10.8106 16.4821 11.1556 16.3698 11.5381 16.4033C11.9207 16.4367 12.2409 16.6072 12.4989 16.9147L19.7839 25.5965L21.9543 23.7753L18.3119 19.4344C18.0539 19.1269 17.9416 18.7819 17.9751 18.3993C18.0085 18.0168 18.179 17.6965 18.4865 17.4385C18.794 17.1805 19.139 17.0682 19.5215 17.1017C19.904 17.1352 20.2243 17.3057 20.4823 17.6131L24.1248 21.9541L26.2952 20.1328L23.5634 16.8771C23.3054 16.5697 23.1931 16.2246 23.2266 15.8421C23.2601 15.4596 23.4305 15.1393 23.738 14.8813C24.0455 14.6233 24.3905 14.511 24.773 14.5445C25.1556 14.5779 25.4759 14.7484 25.7339 15.0559L28.4657 18.3116L30.6362 16.4904L29.7256 15.4051C29.4675 15.0976 29.3553 14.7526 29.3887 14.3701C29.4222 13.9876 29.5927 13.6673 29.9002 13.4093C30.2076 13.1513 30.5527 13.039 30.9352 13.0725C31.3177 13.1059 31.638 13.2764 31.896 13.5839L38.2703 21.1805C39.272 22.3743 39.7044 23.7528 39.5677 25.3163C39.4309 26.8797 38.7656 28.1622 37.5719 29.1639L29.921 35.5837Z"
+            fill="white"
+          />
+        </g>
+        <rect
+          opacity="0.5"
+          x="1"
+          y="3.85474"
+          width="16"
+          height="55"
+          rx="8"
+          fill="white"
+        />
+        <rect
+          opacity="0.5"
+          x="1"
+          y="4"
+          width="16"
+          height="22"
+          rx="8"
+          fill="white"
+        />
+      </Icon>
+    );
+  },
   ChevronRightIcon,
   ChevronLeftIcon,
   ChevronUpIcon,
