+++ src/pages/videos/VideoReels.tsx
@@ -1,39 +1,22 @@
-import {
-  Alert,
-  AlertIcon,
-  AlertTitle,
-  Badge,
-  Box,
-  Center,
-  HStack,
-  IconButton,
-  VStack,
-} from "@chakra-ui/react";
+import { Alert, AlertIcon, AlertTitle, Box, Center } from "@chakra-ui/react";
 import { debounce } from "lodash"; // remove uniqueId
 import React, { memo, useCallback, useEffect, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router-dom";
 import { FixedSizeList as List } from "react-window";
-import PrimaryButton from "../../components/common/button/PrimaryButton";
 import CoinPopover from "../../components/common/cards/CoinPopover";
-import IconByName from "../../components/common/icons/Icon";
-import StarRating from "../../components/common/input/Rating";
 import Layout from "../../components/common/layout/layout";
 import useDeviceSize from "../../components/common/layout/useDeviceSize";
 import Loading from "../../components/common/Loading";
-import CustomHeading from "../../components/common/typography/Heading";
-import SunbirdPlayer from "../../components/players/SunbirdPlayer";
 import * as content from "../../services/content";
 import { callBatch } from "../../services/telemetry";
-import { getSid } from "../../services/utilService";
-import { customLog, handleEvent } from "./utils";
+import { customLog, handleEvent, updateCdataTag } from "./utils";
+import AssessmentPlayer from "./videoReelComponent/AssessmentPlayer";
+import ContentPlayer from "./videoReelComponent/ContentPlayer";
 import CustomSkeleton from "./videoReelComponent/CustomSkeleton";
-import Overlay from "./videoReelComponent/Overlay";
-const VITE_PLAYER_URL = import.meta.env.VITE_PLAYER_URL;
-const VITE_APP_ID = import.meta.env.VITE_APP_ID;
-const VITE_APP_VER = import.meta.env.VITE_APP_VER;
-const VITE_APP_PID = import.meta.env.VITE_APP_PID;
-const TELEMETRYBATCH = import.meta.env.VITE_TELEMETRYBATCH || 5;
+import { TopIcon } from "./videoReelComponent/TopIcon";
+import VideoItem from "./videoReelComponent/VideoItem";
+const TELEMETRYBATCH = import.meta.env.VITE_TELEMETRYBATCH || 20;
 
 /*
 Comment telemetry in all Sunbird players; to undo, find this function and uncomment.
@@ -41,508 +24,6 @@ TelemetrySyncManager.syncFailedBatch();
 TelemetrySyncManager.syncEvents()
 this.sendTelemetry(e)
 */
-
-const contextData = {
-  sid: getSid(),
-  uid: localStorage.getItem("id"),
-  did: localStorage.getItem("did"), // send for ifram data
-  cdata: [
-    {
-      id: localStorage.getItem("grade"),
-      type: "grade",
-    },
-    {
-      id: localStorage.getItem("medium"),
-      type: "medium",
-    },
-    {
-      id: localStorage.getItem("board"),
-      type: "board",
-    },
-    {
-      id: localStorage.getItem("subject"),
-      type: "subject",
-    },
-  ],
-  tags: [
-    {
-      id: localStorage.getItem("grade"),
-      type: "grade",
-    },
-    {
-      id: localStorage.getItem("medium"),
-      type: "medium",
-    },
-    {
-      id: localStorage.getItem("board"),
-      type: "board",
-    },
-    {
-      id: localStorage.getItem("subject"),
-      type: "subject",
-    },
-  ],
-  pdata: {
-    // optional
-    id: VITE_APP_ID, // Producer ID. For ex: For sunbird it would be "portal" or "genie"
-    ver: VITE_APP_VER, // Version of the Application
-    pid: VITE_APP_PID, // Optional. In case the component is distributed, then which instance of that component
-  },
-};
-
-const VideoItem: React.FC<{
-  programID: string | undefined;
-  id: string;
-  qml_id: string;
-  isVisible: boolean;
-  style: React.CSSProperties;
-  refQml?: any;
-  adapter: string;
-  authUser?: any;
-  thumbnailUrl?: string;
-}> = memo(
-  ({
-    id,
-    qml_id,
-    isVisible,
-    adapter,
-    programID,
-    authUser,
-    refQml,
-    style,
-    thumbnailUrl,
-  }) => {
-    const { t } = useTranslation();
-    const [videoEndId, setVideoEndId] = useState<any>({});
-    const [isLoading, setIsLoading] = useState<boolean>(true);
-    const [isQUMLLoading, setIsQUMLLoading] = useState<boolean>(false);
-    const [playerContext, setPlayerContext] = useState<any>(contextData);
-    const { width, height } = useDeviceSize();
-    const [rating, setRating] = useState(0);
-    const [lesson, setLesson] = React.useState<{ mimeType: string }>({
-      mimeType: "",
-    });
-    const [lessonQml, setLessonQml] = React.useState<{
-      mimeType: string;
-      subject?: string | [];
-    }>({
-      mimeType: "",
-    });
-    const [heightPerItem, setHeightPerItem] = useState<{
-      width: number;
-      height: number;
-    }>({ width: 0, height: 0 });
-    const subjectRef = useRef<HTMLDivElement>(null);
-
-    const updateCdataTag = (data: any[]) => {
-      return {
-        ...playerContext,
-        cdata: [...playerContext.cdata, ...data],
-        tags: [...playerContext.tags, ...data],
-      };
-    };
-
-    useEffect(() => {
-      setHeightPerItem({ height: 0, width: 0 });
-    }, [height]);
-
-    useEffect(() => {
-      if (!isVisible) return;
-      const inti = async () => {
-        setIsLoading(true);
-        if (adapter) {
-          let resultData = await content.getOne({
-            id,
-            adapter,
-            type: "course",
-          });
-          if (qml_id) {
-            let qmlResult = await content.getOne({
-              id: qml_id,
-              adapter,
-              type: "assessment",
-            });
-            setLessonQml(qmlResult);
-          }
-          setPlayerContext(
-            updateCdataTag([
-              {
-                id: programID,
-                type: "program",
-              },
-              {
-                id: authUser?.GroupMemberships?.[0]?.School?.udiseCode,
-                type: "school_udise",
-              },
-              {
-                id: authUser?.username,
-                type: "username",
-              },
-              {
-                id: adapter,
-                type: "contentSource",
-              },
-            ])
-          );
-          setLesson(resultData);
-          setIsLoading(false);
-        }
-      };
-      inti();
-    }, [id, isVisible]);
-
-    React.useEffect(() => {
-      const newHandleEvent = async (data: any) => {
-        const result = handleEvent(data);
-        if (!result || !result?.type) return;
-        if (result.iframeId === "assessment") {
-          setVideoEndId((old: any) => ({ ...old, qml_id }));
-        } else if (result.iframeId === "course") {
-          setVideoEndId((old: any) => ({ ...old, id }));
-        }
-      };
-      const handleEventNewItem = (event: any) => {
-        newHandleEvent(event);
-      };
-
-      window.addEventListener("message", handleEventNewItem, false);
-
-      return () => {
-        window.removeEventListener("message", handleEventNewItem);
-      };
-    }, []);
-
-    return (
-      <div
-        style={{
-          ...style,
-          width: "100%",
-          height: "100%",
-          scrollSnapAlign: "start",
-        }}
-      >
-        {!["sunbird", "diksha"].includes(adapter) ? (
-          <Center height="100%" p={4}>
-            <Alert status="error" variant="solid">
-              <AlertIcon />
-              <AlertTitle>
-                {t(
-                  `This (${
-                    adapter || "Unknown Source"
-                  }) content source is not supported, expected source is Diksha and Sunbird`
-                )}
-              </AlertTitle>
-            </Alert>
-          </Center>
-        ) : isVisible && !isLoading ? (
-          <Box>
-            <Box>
-              {videoEndId?.id === id ? (
-                <Overlay {...{ width, height, thumbnailUrl, setVideoEndId }} />
-              ) : (
-                <SunbirdPlayer
-                  LoaderComponent={({ display }: any) => (
-                    <CustomSkeleton display={display} />
-                  )}
-                  {...{ width, height }}
-                  _playerStypeHeight={height}
-                  {...{
-                    ...lesson,
-                    iframeId: "course",
-                  }}
-                  userData={{
-                    firstName: localStorage.getItem("name"),
-                    lastName: "",
-                  }}
-                  public_url={VITE_PLAYER_URL}
-                  adapter={adapter}
-                  playerContext={updateCdataTag([
-                    {
-                      id: qml_id,
-                      type: "question_set",
-                    },
-                    {
-                      id: lesson?.mimeType,
-                      type: "mimeType",
-                    },
-                  ])}
-                  batchsize={TELEMETRYBATCH}
-                />
-              )}
-            </Box>
-            {qml_id && (
-              <VStack bg="red.100">
-                {lessonQml?.subject && (
-                  <HStack
-                    ref={subjectRef}
-                    spacing={1}
-                    visibility={
-                      heightPerItem?.height === 0 ? "hidden" : "visible"
-                    }
-                    opacity={heightPerItem?.height === 0 ? "0" : "1"}
-                    position={"absolute"}
-                    zIndex={20}
-                    right={
-                      heightPerItem?.width === 0
-                        ? "0px"
-                        : `${
-                            heightPerItem?.width -
-                            (subjectRef?.current?.clientWidth || 0)
-                          }px`
-                    }
-                    bottom={
-                      heightPerItem?.height === 0
-                        ? "32px"
-                        : `${heightPerItem?.height - 74}`
-                    }
-                    transition="right 0.5s,bottom 0.5s"
-                    top="auto"
-                  >
-                    {Array.isArray(lessonQml?.subject) ? (
-                      lessonQml.subject.map((sub) => (
-                        <Badge
-                          key={sub}
-                          p="2px"
-                          fontSize="12px"
-                          fontWeight={500}
-                          color="#03627C"
-                          bg="#03627C33"
-                        >
-                          {sub}
-                        </Badge>
-                      ))
-                    ) : (
-                      <Badge
-                        p="2px"
-                        fontSize="12px"
-                        fontWeight={500}
-                        color="#03627C"
-                        bg="#03627C33"
-                      >
-                        {lessonQml?.subject}
-                      </Badge>
-                    )}
-                  </HStack>
-                )}
-                <TopIcon
-                  onClick={() => {
-                    if (heightPerItem?.height === 0) {
-                      setHeightPerItem({
-                        height: height / 2.5,
-                        width: width - 31,
-                      });
-                    } else {
-                      setHeightPerItem({ height: 0, width: 0 });
-                    }
-                    if (!isQUMLLoading) {
-                      setIsQUMLLoading(true);
-                    }
-                  }}
-                  {...(heightPerItem?.height === 0
-                    ? {
-                        rounded: "none",
-                        roundedLeft: "full",
-                      }
-                    : {})}
-                  size={heightPerItem?.height === 0 ? "lg" : "sm"}
-                  _icon={{
-                    width: heightPerItem?.height === 0 ? "100%" : "",
-                    height: "",
-                    fontWeight: "600",
-                    color: "primary.500",
-                  }}
-                  p={heightPerItem?.height === 0 ? "5px 16px" : ""}
-                  icon={
-                    heightPerItem?.height === 0
-                      ? "TakeAQuizIcon"
-                      : "ChevronDownIcon"
-                  }
-                  bg={heightPerItem?.width === 0 ? "white" : "#03627C1A"}
-                  right={heightPerItem?.width === 0 ? "0px" : `${32}px`}
-                  bottom={
-                    heightPerItem?.height === 0
-                      ? "32px"
-                      : `${heightPerItem?.height - 82}`
-                  }
-                  transition="right 0.5s,bottom 0.5s"
-                  top="auto"
-                />
-                {isQUMLLoading &&
-                  (videoEndId?.qml_id === qml_id ? (
-                    <Box
-                      // display={heightPerItem?.height === 0 ? "none" : "block"}
-                      pt={"52px"}
-                      bg={"transparent"}
-                      {...heightPerItem}
-                      {...{
-                        position: "absolute",
-                        bottom: "16px",
-                        transition: "right 0.5s,width 0.5s, height 0.5s",
-                        right: "16px",
-                      }}
-                    >
-                      <VStack
-                        p={heightPerItem?.height === 0 ? "0px" : "4"}
-                        pt={heightPerItem?.height === 0 ? "0px" : "44px"}
-                        bg={"white"}
-                        ref={isVisible ? refQml : false}
-                        rounded={"16px"}
-                        height={"100%"}
-                        textAlign={"center"}
-                        spacing={2}
-                        color={"darkBlue.500"}
-                        overflowY={"scroll"}
-                        justifyContent="center"
-                      >
-                        {rating < 100 ? (
-                          <VStack spacing={2} width={"100%"}>
-                            <CustomHeading
-                              fontSize={"24px"}
-                              fontWeight={"700"}
-                              fontFamily={"Bebas Neue"}
-                              color={"darkBlue.500"}
-                              id="quiz-title"
-                            >
-                              CONGRATULATIONS!
-                            </CustomHeading>
-                            <VStack spacing={1}>
-                              <CustomHeading
-                                id="quiz-subtitle"
-                                fontSize={"14px"}
-                                fontWeight={"500"}
-                                color={"darkBlue.500"}
-                              >
-                                You've completed the quiz!
-                              </CustomHeading>
-
-                              <CustomHeading
-                                fontSize={"14px"}
-                                fontWeight={"500"}
-                                color={"darkBlue.500"}
-                                id="rating-text"
-                                display={"none"}
-                              >
-                                and you have earned
-                                <span
-                                  id="rating-point"
-                                  style={{ padding: "0 5px" }}
-                                >
-                                  Loding.
-                                </span>
-                                coins.
-                              </CustomHeading>
-                            </VStack>
-                            <VStack
-                              bg="white"
-                              id="rating-box"
-                              display={"none"}
-                              position={width < 412 ? "sticky" : "relative"}
-                              bottom={width < 412 ? "0px" : "auto"}
-                              spacing={width < 412 ? "1" : 4}
-                              width={"100%"}
-                            >
-                              <StarRating
-                                value={rating}
-                                onChange={setRating}
-                                hStackProps={{ id: "rating-start", spacing: 4 }}
-                              />
-                              <PrimaryButton
-                                size={width < 412 ? "sm" : "md"}
-                                width={"100%"}
-                                id="rating-button"
-                                isDisabled={rating === 0 ? true : false}
-                                onClick={async () => {
-                                  const result = await content.rateQuiz({
-                                    programId: programID,
-                                    subject: Array.isArray(lessonQml?.subject)
-                                      ? (lessonQml.subject as string[])[0]
-                                      : lessonQml?.subject,
-                                    userId: authUser.userId,
-                                    assessmentId: videoEndId?.qml_id,
-                                    rating: rating,
-                                  });
-                                  if (result) {
-                                    setRating(100);
-                                  }
-                                }}
-                              >
-                                Submit
-                              </PrimaryButton>
-                            </VStack>
-                          </VStack>
-                        ) : (
-                          <VStack spacing={6}>
-                            <IconByName
-                              boxSize={"48px"}
-                              name={"FamilyStarIcon"}
-                              color={"yellow.lightDark"}
-                            />
-                            <VStack spacing={3}>
-                              <CustomHeading
-                                fontSize={"24px"}
-                                fontWeight={"700"}
-                                fontFamily={"Bebas Neue"}
-                                color={"darkBlue.500"}
-                              >
-                                Thank you for your feedback!
-                              </CustomHeading>
-                              <CustomHeading
-                                id="quiz-subtitle"
-                                fontSize={"14px"}
-                                fontWeight={"500"}
-                                color={"darkBlue.500"}
-                              >
-                                Your feedback is valuable to us
-                              </CustomHeading>
-                            </VStack>
-                          </VStack>
-                        )}
-                      </VStack>
-                    </Box>
-                  ) : (
-                    <SunbirdPlayer
-                      LoaderComponent={({ display }: any) => (
-                        <CustomSkeleton display={display} type="assessment" />
-                      )}
-                      forwardedRef={isVisible ? refQml : false}
-                      style={{ border: "none", borderRadius: "16px" }}
-                      _vstack={{
-                        position: "absolute",
-                        bottom: "16px",
-                        transition: "right 0.5s,width 0.5s, height 0.5s",
-                        right: "16px",
-                      }}
-                      {...heightPerItem}
-                      {...{ ...lessonQml, iframeId: "assessment" }}
-                      userData={{
-                        firstName: localStorage.getItem("name"),
-                        lastName: "",
-                      }}
-                      public_url={VITE_PLAYER_URL}
-                      adapter={adapter}
-                      playerContext={updateCdataTag([
-                        {
-                          id,
-                          type: "learning_content",
-                        },
-                        {
-                          id: lessonQml?.mimeType,
-                          type: "mimeType",
-                        },
-                      ])}
-                      batchsize={TELEMETRYBATCH}
-                    />
-                  ))}
-              </VStack>
-            )}
-          </Box>
-        ) : (
-          <CustomSkeleton />
-        )}
-      </div>
-    );
-  }
-);
 
 const VideoReel: React.FC<{
   videos: any[];
@@ -828,44 +309,6 @@ const VideoReel: React.FC<{
 };
 
 export default VideoReel;
-
-const TopIcon: React.FC<{
-  onClick: () => void;
-  icon: any;
-  _icon?: any;
-  right?: string;
-  left?: string;
-  zIndex?: string;
-  top?: string;
-  bottom?: string;
-  transition?: string;
-  rounded?: string;
-  roundedLeft?: string;
-  size?: string;
-  bg?: string;
-  p?: string;
-  _hover?: any;
-  _active?: any;
-}> = ({ onClick, left, icon, _icon, ...props }) => {
-  return (
-    <IconButton
-      aria-label="Go back"
-      icon={<IconByName name={icon} boxSize="2rem" color="white" {..._icon} />}
-      size="mg"
-      variant="ghost"
-      position="absolute"
-      top="16px"
-      left={left}
-      zIndex="20"
-      bg="#FFFFFF26"
-      rounded={"full"}
-      border={"none"}
-      _focus={{ boxShadow: "none", outline: "none" }}
-      onClick={onClick}
-      {...props}
-    />
-  );
-};
 
 const LikeButton: React.FC<any> = ({ playerPayload }) => {
   const [isLiked, setIsLiked] = useState(false);
